fix(eval): respect runtime architecture boundaries
